+++ bin/wp-cli.php
@@ -61,7 +61,7 @@ class Windows_Azure_Storage_CLI extends WP_CLI_Command {
 	 * wp windows-azure-storage containers-list --prefix=demo
 	 */
 	public function list_containers( $args, $assoc_args ) {
-		$assoc_args  = wp_parse_args( $assoc_args, array(
+		$assoc_args = wp_parse_args( $assoc_args, array(
 			'prefix' => '',
 		) );
 
@@ -73,14 +73,16 @@ class Windows_Azure_Storage_CLI extends WP_CLI_Command {
 			'field'  => null
 		);
 
-		$table       = new \WP_CLI\Formatter( $format_args );
-		$containers  = $client->list_containers( $assoc_args['prefix'] );
+		$table      = new \WP_CLI\Formatter( $format_args );
+		$containers = $client->list_containers( $assoc_args['prefix'] );
 
 		if ( is_wp_error( $containers ) ) {
 			return WP_CLI::error( $containers->get_error_message() );
 		}
-
-		$items = $containers->get_all();
+		$items = array();
+		foreach ( $containers as $container ) {
+			$items[] = $container;
+		}
 
 		if ( empty( $items ) ) {
 			return WP_CLI::warning( __( 'No containers found.', MSFT_AZURE_PLUGIN_DOMAIN_NAME ) );
@@ -167,7 +169,7 @@ class Windows_Azure_Storage_CLI extends WP_CLI_Command {
 			'field'  => null
 		);
 
-		$table       = new \WP_CLI\Formatter( $format_args );
+		$table = new \WP_CLI\Formatter( $format_args );
 		$table->display_item( $result );
 	}
 
@@ -210,6 +212,91 @@ class Windows_Azure_Storage_CLI extends WP_CLI_Command {
 				$result
 			)
 		);
+	}
+
+	/**
+	 * Delete blob from container.
+	 *
+	 * @param array $args       Command arguments.
+	 * @param array $assoc_args Command options.
+	 *
+	 * @return void
+	 *
+	 * ## OPTIONS
+	 *
+	 * <container>
+	 * : Container name.
+	 *
+	 * <path>
+	 * : Remote file path.
+	 *
+	 * @subcommand delete-blob
+	 *
+	 * ## EXAMPLE
+	 * wp windows-azure-storage delete-blob testcontainer image1.png
+	 */
+	public function delete_blob( $args, $assoc_args ) {
+		if ( empty( $args ) || count( $args ) < 2 ) {
+			return WP_CLI::error( __( 'Container and remote path must be set.', MSFT_AZURE_PLUGIN_DOMAIN_NAME ) );
+		}
+
+		list( $container, $remote_path ) = $args;
+		$credentials = Windows_Azure_Config_Provider::get_account_credentials();
+		$client      = new Windows_Azure_Rest_Api_Client( $credentials['account_name'], $credentials['account_key'] );
+		$result      = $client->delete_blob( $container, $remote_path );
+
+		if ( is_wp_error( $result ) ) {
+			return WP_CLI::error( $result->get_error_message() );
+		}
+
+		WP_CLI::success(
+			'Blob has been deleted.'
+		);
+	}
+
+	/**
+	 * Get blob properties.
+	 *
+	 * @param array $args       Command arguments.
+	 * @param array $assoc_args Command options.
+	 *
+	 * @return void
+	 *
+	 * ## OPTIONS
+	 *
+	 * <container>
+	 * : Container name.
+	 *
+	 * <remote_path>
+	 * : Blob path.
+	 *
+	 * @subcommand blob-properties
+	 *
+	 * ## EXAMPLE
+	 * wp windows-azure-storage blob-properties testcontainer image1.png
+	 */
+	public function get_blob_properties( $args, $assoc_args ) {
+		if ( count( $args ) < 2 ) {
+			return WP_CLI::error( __( 'Container name and blob path must be set.', MSFT_AZURE_PLUGIN_DOMAIN_NAME ) );
+		}
+
+		list( $container, $remote_path ) = $args;
+		$credentials = Windows_Azure_Config_Provider::get_account_credentials();
+		$client      = new Windows_Azure_Rest_Api_Client( $credentials['account_name'], $credentials['account_key'] );
+		$result      = $client->get_blob_properties( $container, $remote_path );
+
+		if ( is_wp_error( $result ) ) {
+			return WP_CLI::error( $result->get_error_message() );
+		}
+
+		$format_args = array(
+			'format' => 'table',
+			'fields' => array_keys( $result ),
+			'field'  => null
+		);
+
+		$table = new \WP_CLI\Formatter( $format_args );
+		$table->display_item( $result );
 	}
 }